AppleScript droplet for pandoc file conversion
This AppleScript droplet is designed to make and run a shell script
that uses pandoc (the "swiss-army knife" for document conversion) to
convert one or more files whose icons are dragged onto the droplet icon
in the OS X Finder. The script converts multiple input files
recursively, instead of concatenating multiple input files as pandoc
does when run from the command line. We find this droplet to be much
less tedious to use than the command line. To save the script as a
droplet, open the script in Script Editor and save the script with file
format: Application.